add 'cutoff' parameter to function GenTable.
GenTable accept a list of topGOResult. By default, it will sort and order by the first result. If 'cutoff' is provide, GenTable will only keep those terms that have a p-value less than the cutoff.

The 'cutoff' will override the 'topNodes' parameter if the number of terms that pass the cutoff is less than the number of 'topNodes'.